Created a Pathlinker class with all commandline options as properties and add method following methods-
__init__ - Constructor for Pathlinker class
set_network, set_targets and set_sources - setter methods
read_network_file - Reads the network file and returns the network
object
read_nodes_values_file - Reads the nodetypes file and returns a list of
sources and targets
execute - executes the pathlinker algorithm and returns three values :
1) prFinal (node visitation probabilities) will be None if --PageRank
option is not used
2) paths are k shortest paths found using ksp pathfinding algorithm
3) pathgraph is used to prepare the k shortest paths for output to flat files

Created a PathLinkerError class to handle errors throws while executing
Pathlinker algorithm. The value property stores the error message.

Added getKSPGraph method so that it can be reused to get output as text
instead of printing to a file. This method is reused by printKSPGraph to
print the KSPGraph to given file.